+++ src/test/java/org/realityforge/replicant/client/transport/DataLoadActionTest.java
@@ -186,11 +186,6 @@ public class DataLoadActionTest
     }
 
     @Override
-    public void delink()
-    {
-    }
-
-    @Override
     public void invalidate()
     {
     }